refinement: descriptor+branch into own file
